'morePics' added to $newResult array
Some posts have more than 1 picture, and I needed to grab all the pictures for the posts, if it had more. So this is what I've come up with.

I will probably be tweaking this some more to allow us to base64_decode() the images and save them on our system so its not as heavy in the html source. Will update your project when I am able to confirm that is working.